michalPg

+++ api/src/main/java/pl/kolendateam/dadcard/race/entity/SubRace.java
@@ -66,5 +66,5 @@ public class SubRace implements Serializable {
   )
   Set<Region> availableRegions = new HashSet<>();
 
-  byte levelAdjustment;
+  Integer levelAdjustment;
 }
